requirements.txt
accelerate==1.0.1
aiofiles==23.2.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.9
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
anyio==4.5.2
apturl==0.5.2
async-timeout==4.0.3
attrs==24.2.0
bidict==0.23.1
blinker==1.8.2
Brlapi==0.7.0
cached-property==1.5.1
cachetools==5.5.0
certifi==2019.11.28
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.3
command-not-found==0.3
contourpy==1.1.1
cryptography==2.8
cupshelpers==1.0
cycler==0.12.1
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
diffusers==0.30.3
distro==1.9.0
distro-info==0.23+ubuntu1.1
docker==4.1.0
dockerpty==0.4.1
docopt==0.6.2
einops==0.8.0
entrypoints==0.3
exceptiongroup==1.2.2
fastapi==0.115.5
ffmpy==0.4.0
filelock==3.16.1
flask==3.0.3
Flask-SocketIO==5.4.1
fonttools==4.55.0
frozenlist==1.4.1
fsspec==2024.9.0
gitdb==4.0.11
GitPython==3.1.43
gradio==4.44.1
gradio-client==1.3.0
greenlet==3.1.1
gTTS==2.5.4
h11==0.14.0
httpcore==1.0.6
httplib2==0.14.0
httpx==0.27.2
huggingface-hub==0.25.2
idna==2.8
imageio==2.35.1
imageio-ffmpeg==0.5.1
importlib-metadata==8.5.0
importlib-resources==6.4.5
itsdangerous==2.2.0
jinja2==3.1.4
jiter==0.7.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==3.2.0
keyring==18.0.1
kiwisolver==1.4.7
kornia==0.7.3
kornia-rs==0.1.5
langchain==0.2.17
langchain-core==0.2.43
langchain-text-splitters==0.2.4
langsmith==0.1.142
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
louis==3.12.0
macaroonbakery==1.3.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
mdurl==0.1.2
more-itertools==4.2.0
moviepy==1.0.3
mpmath==1.3.0
multidict==6.1.0
narwhals==1.13.5
netifaces==0.10.4
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.77
nvidia-nvtx-cu12==12.1.105
oauthlib==3.1.0
olefile==0.46
openai==1.54.4
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
orjson==3.10.11
packaging==24.1
pandas==2.0.3
pexpect==4.6.0
pillow==10.4.0
proglog==0.1.10
propcache==0.2.0
protobuf==5.28.3
psutil==6.0.0
pyarrow==17.0.0
pycairo==1.16.2
pycparser==2.22
pycups==1.9.73
pydantic==2.9.2
pydantic-core==2.23.4
pydeck==0.9.1
pydub==0.25.1
pygments==2.18.0
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==3.1.4
pyRFC3339==1.1
pyrsistent==0.15.5
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.9.0.post0
python-debian==0.1.36+ubuntu1.1
python-dotenv==1.0.1
python-engineio==4.10.1
python-multipart==0.0.17
python-socketio==5.11.4
pyttsx3==2.98
pytz==2024.2
pyxdg==0.26
PyYAML==5.3.1
regex==2024.9.11
reportlab==3.5.34
requests==2.32.3
requests-toolbelt==1.0.0
requests-unixsocket==0.2.0
rich==13.9.4
ruff==0.7.3
safetensors==0.4.5
scipy==1.10.1
SecretStorage==2.3.1
semantic-version==2.10.0
sentencepiece==0.2.0
shellingham==1.5.4
simple-websocket==1.1.0
simplejson==3.16.0
six==1.14.0
smmap==5.0.1
sniffio==1.3.1
soundfile==0.12.1
spandrel==0.4.0
SQLAlchemy==2.0.36
starlette==0.41.2
streamlit==1.40.1
streamlit-player==0.1.5
sympy==1.13.3
systemd-python==234
tenacity==8.5.0
texttable==1.6.2
tokenizers==0.20.0
toml==0.10.2
tomlkit==0.12.0
torch==2.4.1
torchaudio==2.4.1
torchsde==0.2.6
torchvision==0.19.1
tornado==6.4.1
tqdm==4.66.5
trampoline==0.1.2
transformers==4.45.2
triton==3.0.0
typer==0.13.0
typing-extensions==4.12.2
tzdata==2024.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36
unattended-upgrades==0.1
urllib3==2.2.3
uvicorn==0.32.0
wadllib==1.3.3
watchdog==4.0.2
websocket-client==0.53.0
websockets==12.0
werkzeug==3.0.6
wsproto==1.2.0
xkit==0.0.0
yarl==1.14.0
zipp==3.20.2
